add rake tasks for presym
